Move unix timestamp parser to utils
